# src/main/java/com/example/cst/controller/StudentController.java
package com.example.cst.controller;

import com.example.cst.model.Student;
import com.example.cst.service.StudentService;
import lombok.AllArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@AllArgsConstructor
public class StudentController {

    private final StudentService studentService;

    @GetMapping("/people")
    public ResponseEntity<List<Student>> getStudents() {

        return new ResponseEntity<>(studentService.getStudents(), HttpStatus.OK);
    }

    @PostMapping("/people")
    public ResponseEntity<Student> createStudent(@RequestBody Student student) {

        return new ResponseEntity<>(studentService.createStudent(student), HttpStatus.CREATED);
    }

    @GetMapping("/people/{rut}")
    public ResponseEntity<Student> getStudent(@PathVariable String rut) {

        return new ResponseEntity<>(studentService.getStudentByRut(rut), HttpStatus.OK);
    }

    @PutMapping("/people/{rut}")
    public ResponseEntity<Student> updateStudent(@PathVariable String rut, @RequestBody Student student) {
        return new ResponseEntity<>(studentService.updateStudent(rut, student), HttpStatus.OK);
    }

    @DeleteMapping("/people/{rut}")
    public ResponseEntity<Student> deleteStudent(@PathVariable String rut) {
        studentService.deleteStudent(rut);

        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
    }



}